mcpserver/pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-groq>=1.1.2",
    "mcp-use>=1.7.0",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
    "python-env>=1.0.0",
]
